requirements.txt
openai
langchain
streamlit
python-dotenv
PyPDF2
pandas
langchain_community
langchain-openai
huggingface-hub
langchain-huggingface`

-e .